+++ odata_query/django/__init__.py
@@ -1,2 +1,3 @@
 from .django_q import AstToDjangoQVisitor
 from .django_q_ext import *
+from .shorthand import apply_odata_query