+++ odata_query/sqlalchemy/__init__.py
@@ -1,1 +1,2 @@
+from .shorthand import apply_odata_query
 from .sqlalchemy_clause import AstToSqlAlchemyClauseVisitor