+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- Added shorthands for the most common use cases: Applying an OData filter
+  straight to a Django QuerySet or SQLAlchemy query.
+
 ## [0.4.0] - 2021-05-28
 
 ### Changed